CBD-76: Implementation of the weight function
This commit implements the weight function for the RR*-tree split
algorithm. The final goal function was also implemented. Now
everything should be set to call it from the split function, which
makes a destinction between inner and leaf nodes.

There was also a change in the test files, as somehow Etap didn't
display error messages correctly when some error occured.

Change-Id: Iecc8477202557c32c5249d5aaf7138f09743690c
Reviewed-on: http://review.couchbase.org/16066
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>