CBD-681: Improve tree searching
The tree can now be searched/traversed with a function. This is
needed in order to stop the traversal when a limit parameter is
specified.

Additional functions were added. One to traverse the full tree
without a given bounding box and another one that just returns
the number of elements (not the elements themselves) for a
specific query (or the full tree).

Change-Id: I68a68742f450ae5e72efc9265bbcad138ef6bd64