CBD-732: Add dependecies for geometry (polygon) search
I add the source of the dependencies directly to the repository.
The WKT parser probably won't change at all and the erlgeom
will see little development, which will happen on the official
github repository [1]. The changes will then be copied back to
this repository from time to time.

The reason for not using the dependency management from rebar
is the build infrastructure/tools that are used for Couchbase.

The copies of the wkt and erlgeom repositories contain all files
except the .git directory and the rebar executable.

[1] https://github.com/couchbaselabs/erlgeom

Change-Id: Ifca73499e238479ca35337fb926a49c9e6c0da91