MB-7904: Write body and geometry early
Write the body and the geometry of a KV-node early. This is needed
to make the size calculations easier, that are needed to determine
how many children a node will contain.

This means that the body and the geometry of nodes that will be
inserted are written right at the beginning. All calculations
about splitting etc. will happen afterwards.

Change-Id: Iff48a19ee6903db24270adf9273bd5d7d8a85a08
Reviewed-on: http://review.couchbase.org/25137
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>