MB-8746: First working set view spatial view
With this commit spatial views work with the full stack.
Currently only the initial index building is supported,
no updating yet.

The preferred parameter to query those new spatial views
is with using the "range" parameter. There's "start_range"
and "end_range". Both take an array of numbers that
correspond to every dimension of the index. The "start_range"
is for the lower bound, the "end_range" for the upper bound.

Change-Id: If924e316b051eb85ff89c9bce68812634bc7d11a
Reviewed-on: http://review.couchbase.org/28074
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>