MB-12249: Make vtree unit tests pass again
The on-disk file format was changed. The body as well as the geometry
are stored as part of the value and not separately. Hence the
`vtree_io:write_kvnode_external/2` and `vtree_io:read_kvnode_external/2`
functions don't exist anymore.

Due to this change the size of a single kv_node increased, hence the
kv_chunk_threshold needed to be increased for some unit tests.

Change-Id: Id031e2601f0b40cb16caada5bd029c4f047fd976
Reviewed-on: http://review.couchbase.org/41767
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>