MB-15975: Always find a split candidate
There was a case where no split candidate was found. It happens when
there are large nodes (e.g. due to large geometries) where the splitting
rules need to be relaxed.

In this particular case it happens when there are two nodes where the
minimum fill rule can't be achieved for both nodes. It is then ignored.
Instead the first partition of the split candidate is filled up to the
maximum fill rate.

If the last node in a list of nodes leads to this overflow, it will still
be part of the first partition. This would lead to an empty second partition.

The fix is to always make sure the second partition has at least one node.

Change-Id: I1971306f20fbc0afe917b8754bd8e430f43a58c1
Reviewed-on: http://review.couchbase.org/54222
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>